feat:calculate percentage and add resources for products and productImages

+++ database/migrations/2023_07_19_121932_add_discount_to_products.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->decimal('discount', 8, 2)->default(0.00)->after('price');
+            $table->integer('discount', )->default(0)->after('price');
         });
     }
